+++ concordium-node/src/plugins/mod.rs
@@ -1,5 +1,3 @@
 //! Client plugins.
 
 pub mod consensus;
-#[cfg(feature = "staging_net")]
-pub mod staging_net;