+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Update dependencies, code cleanup, and removal of the staging_net feature. No functional changes.
+
 ## concordium-node 1.0.1
 
 - Check that baker keys are consistent (private key matches the public one) on startup.